+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/PrivilegesEvaluatorResponse.java
@@ -1,0 +1,52 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.identity;
+
+/**
+ * Response for privilege evaluation that determines if a request to perform an action is permitted
+ *
+ * @opensearch.experimental
+ */
+public class PrivilegesEvaluatorResponse {
+    boolean allowed = false;
+    PrivilegesEvaluatorResponseState state = PrivilegesEvaluatorResponseState.PENDING;
+
+    public boolean isAllowed() {
+        return allowed;
+    }
+
+    public PrivilegesEvaluatorResponse markComplete() {
+        this.state = PrivilegesEvaluatorResponseState.COMPLETE;
+        return this;
+    }
+
+    public PrivilegesEvaluatorResponse markPending() {
+        this.state = PrivilegesEvaluatorResponseState.PENDING;
+        return this;
+    }
+
+    public boolean isComplete() {
+        return this.state == PrivilegesEvaluatorResponseState.COMPLETE;
+    }
+
+    public boolean isPending() {
+        return this.state == PrivilegesEvaluatorResponseState.PENDING;
+    }
+
+    @Override
+    public String toString() {
+        return "PrivEvalResponse [allowed=" + allowed + "]";
+    }
+
+    public static enum PrivilegesEvaluatorResponseState {
+        PENDING,
+        COMPLETE;
+    }
+
+}